feat: implement content-addressed storage. (#7)
* feat: implement content-addressed storage.

This commit implements content-addressed storage of response bodies
in the default storage implementation.

It removes the file advisory lock on the response body file which
will allow writers to not have to wait for readers of the response
body to finish reading the bodies.

The implementation calculates a Blake3 hash of the content as it
is streamed to a temporary file; it then atomically renames the
temporary file to its location in storage.

As an added benefit, this deduplicates the storage of identical
response bodies.

Fixes #6

* chore: code review feedback.

Make the `tempfile` dependency a workspace dependency.