fix: fix error encountered with Azure Storage.
This commit fixes the handling of a 304 response where the server is not
responding with the correct headers (e.g. Azure Storage).

A 304 response for a resource should have the same headers as a 200 response
for the same resource; when those headers are missing, we error saying that the
server said it was unmodified, but the cache says the body needs updating.

Instead of returning an error, we now serve up the stale cached body as the
server says it was unmodified. However, as we don't update the stored policy
object, a future request for the same resource will always preform revalidation
until the server responds with a 200 or a correctly implemented 304.

This commit includes an update to the dependencies and a reformatting of
`Cargo.toml` based on `cargo sort`.