refactor: refactor cache body implementation to stream to cache. (#11)
* fix: allow partial content responses to be cached.

* refactor: refactor cache body implementation to stream to cache.

This commit refactors the cache body implementation so that the response body
is streamed to the cache rather than written to the cache first and then
returning a body which reads from the cache.

* refactor: delete lock file implementation.

This commit deletes the lock file implementation in favor of using the one from
the stdlib as of Rust 1.89.0.

* chore: fix missing doc.

* chore: use a workspace Rust version.

* chore: code review feedback.

* chore: delete additional files related to file locking.