fix: add back `x-cache-digest` header. (#13)
The previous PR removed the `x-cache-digest` header in all responses.

It _should_ have limited the removal to just the times when the cached body
isn't being served from the cache. The other instances where we are serving the
body from the cache can set the header.

Additionally, this updates the `http-cache-semantics` dependency to go back to
the upstream crate that doesn't support 206 responses.

Given the complexity and narrow scope of supporting very specific ranged requests,
it wasn't worth the trouble of implementing it.