feat: add a "revalidation hook" to `Cache` to allow for recalculation of authentication. (#15)
* feat: add a "revalidation hook" to `Cache` to allow for recalculation of authentication.

This commit adds a `with_revalidation_hook` method to `Cache` that
allows users to control the headers sent as part of a cache revalidation
request.

This is needed for Azure Storage where they inexplicably use conditional
request headers like `If-None-Match` as part of the authentication
signature, which breaks transparent caching middleware like this one.

With the hook, a new `Authorization` header can be calculated and then
overridden for the request to revalidate.

* chore: update CHANGELOG.

* chore: update debug message.

* chore: update README.